+++ extensions/Zay-ES-Net/release-notes.md
@@ -1,10 +1,12 @@
-Zay-ES-Net v1.2.2 (unreleased)
+Zay-ES-Net v1.3.0 (unreleased)
 ------------------
 * RemoteEntityData.completeEntity() modified to check the components against
     the filter to fix the case where two entity sets share the same components
     but not the same filter.  See PR #18
 * Fix NPE in RemoteStringIndex when asking for a non-existant string. See Issue: #14
 * Set sourceCompatibility to 1.7 and turned on detailed 'unchecked' warnings
+* Fixed a ton of 'unchecked' related issues.  Many of the generic method signatures
+    have changed.
 * Some javadoc formatting fixes.